proxy/config: actually move config and fix windows test
